Move `SyncLeapValidationMetaData` to separate file
